requirements.txt
asgiref==3.11.1
Django==4.2
django-cors-headers==3.13.0
django-environ==0.13.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
packaging==26.0
pillow==12.1.1
psycopg==3.3.0
psycopg-binary==3.3.0
PyJWT==2.12.1
pytz==2026.1.post1
qrcode==8.2
setuptools==82.0.1
sqlparse==0.5.5
wheel==0.46.3
